refac(math): move `curve_type.h` to `geometry` directory
