--- tachyon/math/elliptic_curves/curve_type.h
@@ -1,13 +1,0 @@
-#ifndef TACHYON_MATH_ELLIPTIC_CURVES_CURVE_TYPE_H_
-#define TACHYON_MATH_ELLIPTIC_CURVES_CURVE_TYPE_H_
-
-namespace tachyon::math {
-
-enum class CurveType {
-  kShortWeierstrass,
-  kTwistedEdwards,
-};
-
-}  // namespace tachyon::math
-
-#endif  // TACHYON_MATH_ELLIPTIC_CURVES_CURVE_TYPE_H_
