+++ tachyon/math/elliptic_curves/short_weierstrass/jacobian_point.h
@@ -15,10 +15,10 @@
 #include "tachyon/base/logging.h"
 #include "tachyon/math/base/groups.h"
 #include "tachyon/math/elliptic_curves/affine_point.h"
-#include "tachyon/math/elliptic_curves/curve_type.h"
 #include "tachyon/math/elliptic_curves/jacobian_point.h"
 #include "tachyon/math/elliptic_curves/point_xyzz.h"
 #include "tachyon/math/elliptic_curves/projective_point.h"
+#include "tachyon/math/geometry/curve_type.h"
 #include "tachyon/math/geometry/point3.h"
 
 namespace tachyon {
