+++ tachyon/math/elliptic_curves/short_weierstrass/sw_curve.h
@@ -4,12 +4,12 @@
 #include <utility>
 
 #include "tachyon/math/elliptic_curves/affine_point.h"
-#include "tachyon/math/elliptic_curves/curve_type.h"
 #include "tachyon/math/elliptic_curves/jacobian_point.h"
 #include "tachyon/math/elliptic_curves/point_conversions.h"
 #include "tachyon/math/elliptic_curves/point_xyzz.h"
 #include "tachyon/math/elliptic_curves/projective_point.h"
 #include "tachyon/math/elliptic_curves/short_weierstrass/sw_curve_traits_forward.h"
+#include "tachyon/math/geometry/curve_type.h"
 
 namespace tachyon::math {
