+++ tachyon/math/geometry/curve_type.h
@@ -1,0 +1,13 @@
+#ifndef TACHYON_MATH_GEOMETRY_CURVE_TYPE_H_
+#define TACHYON_MATH_GEOMETRY_CURVE_TYPE_H_
+
+namespace tachyon::math {
+
+enum class CurveType {
+  kShortWeierstrass,
+  kTwistedEdwards,
+};
+
+}  // namespace tachyon::math
+
+#endif  // TACHYON_MATH_GEOMETRY_CURVE_TYPE_H_
